doctorbox-server/requirements.txt
--no-index
--find-links vendor/

psycopg2==2.7.3.2
django==2.0
djangorestframework==3.7.3
pytz==2017.3
gunicorn==19.7.1
whitenoise==3.3.1
dj-database-url==0.4.2
